Add --companion to directly talk to a single companion without a connect
Summary:
This will allow overriding the hostname:port of any companion to connect to.

It's intentionally using HOST:PORT so as to make the syntax terser (no need for --companion-port and --companion-host) and to make it harder to avoid missing specifying both.

This means that all invocations can override to talk directly to a companion. This means that for automated scenarios you don't even need to `idb connect` and can avoid the state management that it entails.

I've also defined this at the root parser on purpose. This makes it easier to prefix all commands with this string and avoid having to know where the `--` options need to be interespersed. This makes iteraton more rapid when switching between commands since you can just mess with the tail of the command.

Since previous diffs have removed the management client from a host of commands, we're also now able to substitute this with a direct client, without any of the management functions. This means we can also bypass the state management that occurs in the grpc management client and just send the requests directly.

Reviewed By: c-ryan747

Differential Revision: D19448338

fbshipit-source-id: 71afa87b05507fa2ccb80322a534cd7955b2bf54