Use SSL-based Send/Recieve Messages for FBDeviceLinkClient
Summary:
Prior to iOS 13, the usage of SSL was very loosely enforced. This meant that we could send data in the plain over the wire and everything would be ok.

However, that's now not the case and we were sending and recieving data on the socket without any SSL-based wrapping. This can be seen based on the `SSLContext` variable on an `AMDServiceConnectionRef`.

This is the solution, which involves using the `AMDServiceConnectionSend` and `AMDServiceConnectionRecieve`. I can't see an easy way of using the SSLContext to convert bytes as they are encrypted, but these functions will do `SSL_read` and `SSL_write` for us if there is a secure context. This does mean that we end up using synchronous read and write functions, but if anything this makes the solution easier to follow. It doesn't mean we can use GCD for non-blocking IO at all times, but in reality we can do this on background queues so there's no loss in terms of asynchrony. The flow also becomes slightly easier to follow as we don't need "Future wrapping helpers" in order to have something that resembles "prodedural IO". Instead we just perform the reads and writes on a background queue.

Because we end up reading large payloads with screenshots, I did discover that the Send/Recieve calls don't represent "less bytes written than available" as an error condition. There's an upper limit on the send/receive side so we have to chunk until there's no more bytes to send/recieve, or an error condition is met. This means that we need to chunk within a loop. `NSData` has the `subdataWithRange:` method for doing this. If the input bytes are immutable then `dataWithRange:` doesn't copy the underlying data so there's little cost for doing so.

Reviewed By: zeyadsalloum

Differential Revision: D19468885

fbshipit-source-id: 788658fd54d25a07b2d17dcca7e6c4da66d85c83