Use local knowledge of companion state in order to report accurate companion info
Summary:
This data was previously coming exclusively from what the companion is reporting.

Instead we need this to reflect the knowledge local to the execution of the idb client. This is because a remote companion may have a different notion of it's own port if there is port-forwarding occuring. We should report the port that the companion was registered with **locally** rather than the remote notion

Where we have the "knowledge" about local companions, we use that instead of what is reported by GRPC.

Reviewed By: zeyadsalloum

Differential Revision: D19496367

fbshipit-source-id: 4bb9c770bf894ae87487653a4ef620c17b3e1932