Log error messages to stderr
Summary:
When --json is provided it's reasonable to expect that the stdout of idb will always be parsable as json. At the moment this is not true with exceptions as these are printed as raw text to stdout.

Changing this to be on stderr makes more sense and should solve this issue

Reviewed By: lawrencelomax

Differential Revision: D19640481

fbshipit-source-id: 6677b053a48ccb873cba976966c091d14d41c499