Streamed Output
Summary:
In April 2017 I had to revert D4969929 for producing inconsistent test output for providing streamed output.

One of the reasons for making this change is it will allow us to see incremental reported output in `fbxctest` instead of having to wait for the entire test execution to finish before providing this output. Right now if the test execution hangs or otherwise times-out, we won't get any output other than stderr, meaning it's much harder to diagnose and figure out what to do with hanging tests.

The reason for this revert was the incorrect early-termination reporting, which is now fixed and has a corresponding unit test for it.

run_all_tests

Reviewed By: zeyadsalloum

Differential Revision: D19662660

fbshipit-source-id: 922bd6dc97ad676c076ab71ec0f84295c392fb3c