FBAMDeviceManager should update old-references that wrap AMDeviceRef
Summary:
This is coming up in `idb` since it has the pattern of holding onto a `FBDevice` reference, which transitively strongly-references `FBAMDevice`.

Due to this behaviour, if the device is disconnected at any point and re-connected the `AMDeviceRef` in the `idb`-help `FBAMDevice` will no longer be valid! As shown in the existing implementation of the manager, we should *always* use the latest reference to an `AMDeviceRef` irrespective of whether this is for a newly-connected device, or a new object that supercedes the old one.

As such we should ensure that for anyone holding a reference to an `FBAMDevice` elswehere in the code, it should be updated with the new `AMDeviceRef` irrespective of whether this is from a disconnected -> connected transition. This is true in the case of `idb`.

To fulfill this invariant, we maintain two dictionaries:

- A dictionary that strongly references it's values, for maintaining an index of attached devices. This will be cleared when a device is disconnected.
- A map table that weakly references it's values, for maintaining an index of all devices that are either attached or referenced. This will clear it's values when they are no longer strongly-refernced elsewhere. It means we can re-use the same `FBAMDevice` instance, rather than creating a new one if someone is already referencing it.

There's some additional logging to make what is going on clearer

Reviewed By: zeyadsalloum

Differential Revision: D19856310

fbshipit-source-id: 1f082e04a89ac1e953feb6379782c9bedd1e02b9