Don't re-alias loaded into local state for direct_companion_manager
Summary:
Each of the cases that use the companions go to disk each time. This is justifiable since in a multi-process environment, the list of companions could be modified behind the back of a process.

Given that this is the case, there's no need to re-alias this into an property all of the time. We can instead make this a product of internal functions that is only locally valid and expose this in a CM.

The CM is then responsible for seeing whether the companion list is modified and if it is, flushing it to disk. This also allows us to optimize the opens so that we only need to do so once, without locking twice.

Reviewed By: c-ryan747

Differential Revision: D19855628

fbshipit-source-id: fcb129addf2373f0bfee733ec4ab1f4403540da6