Improve unit tests for the direct companion manager
Summary:
This adds a series of tests that cover the public API of the companion manager.

I've also made it possible to iterate over the same test with multiple types of file. This covers the case that the file exists first or that it doesn't. The sequences are also longer to test compound behaviour.

Additionally, the return type of remove companion now provides the companions that were removed to make it easier to test

Reviewed By: c-ryan747

Differential Revision: D20007095

fbshipit-source-id: 020b62718b0ac5ea8e453b99e830a77af996059d