Replicate the in-memory behaviour of the daemon by having it clear companions
Summary:
This is a temporary fix whilst we migrate away from the daemon.

In order to preserve some of the old in-memory beaviour of the daemon, we should ensure that the state of companions is cleared on setup and teardown. This will ensure that any state persisting over CI runs is eliminated

Reviewed By: c-ryan747

Differential Revision: D20028988

fbshipit-source-id: 864901d3d733f7a60d115af3d9a41cde80939065