Adding the same udid twice should result in replacement
Summary:
This case was not previously used and due to the migration towards lists we accidentally lost this behaviour.

Instead of changing internally back to maps, it's easier to add a test for this and fix the behaviour internally.

This also adds showing the evicted item in the return so that we can also assert on this.

Reviewed By: c-ryan747

Differential Revision: D20028582

fbshipit-source-id: 56ee639903d888a7df4648473d27c5c0b9550ab8