Populate failure info for crashes
Summary:
We realized in https://fb.workplace.com/groups/ctp.integrations/permalink/555817105005658/ that for the last couple of months (at least), many test failures in fbobjc have no accompanying stack trace. We debugged it in T62797682 back to IDB itself, and found that when the `idb xctest run` call crashes (usually due to an ASSERT failing or something similar). While we have the logs and the top of the stack available in Sandcastle, it not showing up in TestX made it difficult for engineers to fix their tests. In this diff, I'm addressing it by regex matching the output against what we know the failed assert statement looks like in fbobjc (`Assertion failed: ..., function ..., file ..., line ...`) and populating the failureInfo thereafter. That being said, I've never hacked on idb before so am more than happy to alter the solution if there's a better way to handle this.
`

Reviewed By: lawrencelomax

Differential Revision: D20277073

fbshipit-source-id: 87fb5d615d31d45068c31714caf0c6386363cb96