Add vcard support in add-media
Summary:
Adds support for vcards now that this is supported in the Simulator.

There's more refinement to do here later, but for the meantime this allows for contact updating. This now simplifies the flow somewhat by inlining the work here. I've also had to add support for NSURL based selection so that we don't have to keep going back and forth between urls and strings, we can just keep using strings

Reviewed By: zeyadsalloum

Differential Revision: D20423656

fbshipit-source-id: 4947ede092a9ce582f2d00cb955df627827ff669