log the missing key and existing keys in the dicts of Xcode 11 result bundle
Summary:
In rare cases, IDB might hang while parsing the result bundle. It seems that it always happened after:
```
idb_companion[27078:17072313] *** Assertion failure in id readFromDict(NSDictionary *__strong, NSString *__strong, __unsafe_unretained Class)()
```

I suspect that some fields in the result bundle are dynamic (they might be absent based on the test execution). We did not handle the `nil` case very well. Unfortunately, we do not have enough logs to tell which fields were missing.

When an expected key is missing, we only log the keys of the dicts in the result bundle instead of the entire dicts because the dicts are huge. The entire dict would be really helpful but the keys should give us enough info to tell what might go wrong.

Differential Revision: D20506117

fbshipit-source-id: 2265f3b68701244f786a95358f4f9c5d3ece82b9