Add Companion pruning on non-management command
Summary:
We have this functionality in the management related commands, but now we need to include them in the main set of commands. This uses the connect call failed wrapped exception but this can be expanded over time depending on the use-case

This will help with scenarios that an `idb kill` or `idb list-targets` is needed to clear a dead companion away.

Reviewed By: c-ryan747

Differential Revision: D20667515

fbshipit-source-id: 4df5eb3ce926e9c3c96721fa459a7b0c7a11e77b