Spawned companions should not get SIGINT from parent
Summary:
This is a problem for long-lived commands like `idb log` where sending a `SIGINT` to the python CLI will also deliver the `SIGINT` to the companion as it's spawned as a subrpocess.

For commands that terminate like `list-apps` this isn't an issue, since the process is immediately re-parented when the cli dies.

Reviewed By: zeyadsalloum

Differential Revision: D20667517

fbshipit-source-id: 4ac78a603ff5598abc528d33623c5beeb82e9f70