Allow override wait time for video recording via environment.
Summary:
As title. For tests running longer than 2 minutes we oftenly saw
fbxctest failed to record a video, and the reason probably because video
recording process didn't exit within 10 seconds after we send SIGINT. It's
probably just because video encoding is much slower on older CPU and we need
to wait longer.

Also added some logging so we know when this happened and we have to use SIGKILL
to kill the process.

Differential Revision: D20904862

fbshipit-source-id: 1916988a4c394264a3b5d53b0057346952917709