Positional UDID in Boot
Summary:
This brings the UDID usage inline with the other companion related commands.

It also introduces a new base parser that will then be inherited by the other udid targeted commands. This makes it clearer when the UDID is or isn't required by basing this on any value that's defaulted.

Reviewed By: c-ryan747

Differential Revision: D20556816

fbshipit-source-id: 4f0809f8b67e62209e78b967ba247a0cafcaeddb