parse new activity data
Summary:
Adding to the data about activities that we return from idb to return pretty much everything that xcode does. I'm returning them in a stacked fashion, so that a sub activity is contained within a top level activity. I figure this is a more useful form to consume the information in.

When this option is enabled execution does become slower as xctest will pause to take screenshots etc, but thats an acceptable sacrifice for the people that want this extra data

Reviewed By: lawrencelomax

Differential Revision: D20986561

fbshipit-source-id: 909b6aaab93665537d24c212bc8c15ba57bfb30c