Ensure that we report failure of the test plan in app-test execution
Summary:
The terminating condition for idb's handler is in the reporter. It first waits for the test execution to be done and then for the reporting to be done. This is so that we don't end up writing to a dangling pointer in the handler with any async writing.

What we were observing with a crashing app test was that a crash would occur, so the operation failed. We then wait for the reporting to end, which will never terminate, since it hasn't been handed a terminating condition in it's callback. This ensures that we pass adequate error information during execution towards the reporter. This means that the reporting is then terminated

Differential Revision: D21276884

fbshipit-source-id: 953100c98d0badd5d6ce0e3ff90c9ec67f9954d3