Don't crash with an unpaired device
Summary:
At the moment when the idb companion is run in notify mode with an unpaired device it'll crash trying to serialise that device. This is due to the amdevice object not being created and so some of these properties are null.

As a fix(/hack?) I'm replacing these null values with `unknown`

Reviewed By: zeyadsalloum

Differential Revision: D21301699

fbshipit-source-id: 43b18c6852e2126bd44c18685bab469c82a30653