Setup signal handlers before fetching companion futures
Summary:
I'm currently seeing a race condition on the consumer-side of idb that shows the following:

```
.2425:--headless 1 --boot 1BA654B9-C380-4200-996A-82ABFBD44FA7 - INFO - Stopping process with SIGTERM, waiting 30 seconds
.2425:--headless 1 --boot 1BA654B9-C380-4200-996A-82ABFBD44FA7 - INFO - Process has exited after SIGTERM with -15
```

This will occur when the process is signalled but a signal handler has not been setup. The reason for this is as follows:
1) The py-side of idb launches the companion and waits for stdoiut to contain the relevant output
2) The companion side boots, then writes to stdout
3) The py-side has read the output and then yields, since the simulator is now booted
4) The py-side may then immediately exit the contextmanager, which means that the process is then SIGTERM'd
5) The companion side has not had the opportunity to install the signal handlers, so the process is immediately killed. A SIGTERM with no signal handler is essentially a SIGKILL.

This explains that the process exits with a signal (negative exit code) instead of the cancellation running first. This will mainly only happen when the `boot_headless` CM is exited as soon as it is entered.

Reviewed By: c-ryan747

Differential Revision: D21401615

fbshipit-source-id: a3ad59aa8bf061fea3362a0a3ebcd9facab81099