Add logging of application output
Summary:
This can be very helpful when dealing with crashes. We just need to thread through the right classes to pass this back to the logger that is then used in the reporting mechanism.

These being correct is an invariant, since the logger must be a consumer and is not user-supplied

Reviewed By: c-ryan747

Differential Revision: D21301378

fbshipit-source-id: 9b4d64d59fed65b7328079c7cf7022624fe1e353