Add stdout based reporting of target updates to companion
Summary:
Adds an additional mode to `--notify` in the companion.

If "stdout" is passed to this argument, then the companion will write updates to stdout instead of the given file path. This is desirable for tying the lifecycle of the companion process to that of the caller. It also means that the caller can get updates as soon as they appear. When the stdout pipe of the consumer is closed then this will also close the companion process.

The output is essentially the same, instead we have each update separated by a newline, but the format is still the same JSON overall.

Reviewed By: fejese

Differential Revision: D21595224

fbshipit-source-id: 6863af5352fde7bc4c031656c760219312210be3