Attempt to use XCTestDevices directory
Summary:
Since recent-ish versions of Xcode, Xcode now has it's own Device Set which it uses for management of Simulators. This is distinct from the Default Device Set and likely exists for similar reasons to why `fbxctest` uses a separate device set; we don't want it to interfere with the Simulators that a user is messing around with in an IDE.

Since a more recent-ish version of Xcode (I think 11.4) CoreSimulatorService will now automatically create loads of Simulators for new Device Sets. This can race with `fbxctest` using it for a test invocation and lead to failures as well as the abandonment of device sets that accumilate on the filesystem. Ideally we'd be able to disable this, but I haven't yet found a way to do this.

Additionally, we've seen cases of `CoreSimulatorService` still logging about Simulator sets that have been long-deleted. I suspect that it doesn't have particularly great handling of Simulator Sets being deleted from the filesystem behind it's back. I've also yet to find an API that would allow us to tell `CoreSimulatorService` that we're no longer interested in a particular device set and want it gone.

Either way, the best comporomise is to use this device set that is "intended for automation". We'll fallback to the old behaviour if it's not there

Reviewed By: rlopezdiez

Differential Revision: D21765179

fbshipit-source-id: 883bbb7d566ba72cdc80379740c81c371e0b4f03