Remove CompanionInfo extraction from JSON parser
Summary:
This isn't needed since the only callsites that use this are not emitting CompanionInfo.

The only place where this merging is relevant is on the management client's `list_targets` where the companion infos are merged with the `TargetDescription`s from the output. This is fully external from this point

Reviewed By: c-ryan747

Differential Revision: D21786468

fbshipit-source-id: 9067880157d0b474e422bb970b37393fa344a683