Add -[FBApplicationDataCommands fileCommandsForRootFilesystem]
Summary: We'll want to re-name this commands class later on (it's no longer app-centric) but this shows why we want to do this. I can now manipulate files on a device or simulator relative to the root. The implementation is all there, we just need to provide the right information to the file manipulations and it all works together.

Reviewed By: c-ryan747

Differential Revision: D21685770

fbshipit-source-id: b69363ea9e1a00df30b22a132d2c924db1bcf15c