Add list_targets to companion
Summary:
This is very much not the stateful listing where we can have external companions. This is used for understanding what targets there are of a given type for other kinds of automation.

This is useful when you want to be able to find out in a fast way what devices are connected to the host

Reviewed By: c-ryan747

Differential Revision: D21686365

fbshipit-source-id: 83c58d238227bbdcb244cb2bac887789dca95657