Extract Informational properties to base protocol
Summary:
This is going to be useful for reporting "Restorable Devices". These are devices whereby we have limited information and ability to interact with them.

Since these exist on the base protocol, we can instead just pull them to a new base and then extend. This then uses protocol composition as we do with the rest of it

Reviewed By: fejese

Differential Revision: D21902123

fbshipit-source-id: 81eec125860e837f9266296d2daaf35fb5fdc0f2