Add FBAMRestorableDevice
Summary:
`AMDevice` is only used with booted and working devices. This means that device discovery for non-booted devices in idb is non-existence.

However tools such as `cfgutil` expose the ability to list devices that are in DFU or recovery mode. Fortunately the APIs for finding out about these devices aren't too hard to deal with. The APIs themselves also operate very similarly with `AMDevice` instances. This means that the pulling out of `FBiOSTargetInfo` as well as `FBDeviceManager` can now be used within an `AMRestorableDeviceRef`.

Reviewed By: c-ryan747

Differential Revision: D21881812

fbshipit-source-id: 37f5cd8a38633f62142c6ce243567a1fa47d90a4