Explicit naming of events in FBiOSTargetSetDelegate
Summary:
There's some pretty odd behaviour in the way that updater works in order to mutate the set of values that are known to the updater.

Instead we can make this all explicit and it becomes far easier to reason about what is going on.

This is also now modeled on the `uniqueIdentifier` property whuch uses ECIDs on iOS Devices. This is just internally used for keying and does not change the rendering of udids in the output.

Reviewed By: c-ryan747

Differential Revision: D21903928

fbshipit-source-id: 0483780b1aa8c779d750552f9f0ece3b6a782468