Avoid returning invalid AMDevice instances
Summary:
We don't need to make this a side-effect of the setter and therefore avoid the "Utilize Device" problem. These will be called before we've passed the AMDeviceRef to the device.

Avoiding deferring here also resolves a problem with empty values caused by the `CopyValues` failing silently. This way we will get the values ahead of time and only expose this if they are present

This will eventually enable hiding unpaired devices

Reviewed By: fejese

Differential Revision: D21996046

fbshipit-source-id: 240c3ffe67d03c3b18b757d37e76662d2ec82a8a