Fix companions using SO_REUSEPORT
Summary:
This is a socket option that we have to disable since it is enabled by default in the GRPC librart. We clearly don't want two companions binding on the same port as this can cause a lot of confusion.

Multiple companions for the same target are still fine, we just don't want non-determinism by having two companions bind on the same port ever

Reviewed By: c-ryan747

Differential Revision: D22067142

fbshipit-source-id: 686b0996be437f654d59bae5b5714bb8b73a0824