Fix problems with libdispatch in fifo reading/writing
Summary:
There's a problem in the way that using fifos for stdout/stderr reading within simulator application launches works, this bug also appears to be replicated in simctl in a slightly different way.

In order that we can read the output of an Application process as a data stream, we need to do this via a fifo. The reason for this is that the `CoreSimulator` APIs for launching an Application have `stdout` and `stderr` options that only allow supplying a file path. This means that a file descriptor cannot be passed instead as would be typically done with process output piping. Based on what I can see, SpringBoard (or whatever eventually launches the app on iOS) will connect stdin/stderr correctly as file descriptors. This can be seen by using `lsof` on the process.

```
51603 lawrencelomax    0r    CHR                3,2        0t0                 313 /dev/null
51603 lawrencelomax    1u   FIFO                           0t0           224858910 /private/var/folders/9b/zp4s_s0j0ml66ss6mvxt352x5gyrl5/T/3CDB816D-0960-4D01-AF78-F6F09073C1EE
51603 lawrencelomax    2u   FIFO                       0t12600           224858911 /private/var/folders/9b/zp4s_s0j0ml66ss6mvxt352x5gyrl5/T/3BD6 (https://github.com/facebook/idb/commit/dc1b5c6fcb1380328228d0c58c85a2cbc1354f97)A5EB-C4AF-4322-8E1B-3699F5C1E833
```

This is implemented the same in simctl, which can be seen by launching an app with `--console` and then using `lsof -p` on it's pid:

```
$ xcrun simctl launch --console booted com.apple.Preferences
com.apple.Preferences: 41013
# In another shell
$ lsof -p 41013
...
Preferenc 41013 lawrencelomax    0r   CHR                3,2        0t0                 313 /dev/null
Preferenc 41013 lawrencelomax    1u  FIFO                           0t0           225536069 /Users/lawrencelomax/Library/Developer/CoreSimulator/Devices/E73A9EDB-E1F7-4B52-B26D-5EB77022F3CE/data/tmp/launch_console-41009-1
Preferenc 41013 lawrencelomax    2u  FIFO                           0t0           225536070 /Users/lawrencelomax/Library/Developer/CoreSimulator/Devices/E73A9EDB-E1F7-4B52-B26D-5EB77022F3CE/data/tmp/launch_console-41009-2
```

I can't remeber exactly when this was implemented in simctl exactly, but I recall that pty output was supported first, followed by stdout/stderr passing. The way that this is implemented in `simctl` is similar, it doesn't use `dispatch_io` like this does, but it does use libdispatch to provide read notifications on the read end of the FIFO by using `DISPATCH_SOURCE_TYPE_READ` and subsequent reads of the underlying file descriptor. Fundementally this is very similar to how a dispatch_io channel does async file reading (with reading semantics on top), all of which is implemented using kqueues on macOS.

All of this means that the current implementation looks sound, however it is possible to reproduce an issue with fifo reading whereby the writing process (the application) blocks on a write to stdout/stderr. I don't have a particularly consistent repo for this currently, but I do have an application that does reproduce this. It is happening within an NSLog statement and subsequent write of the stderr fd. Since this happens on the main thread, this is game over for the process.

Quite why this is happening, I'm not sure. My expectation is that once the read of these fds is setup, that everything should work as normal. We do see a significant amount of output flowing over the fifo and I can't seem to see that this a precise amount of data which would suggest some blocked writing. I do suspect this has more to do with the amount of data writen within a single write, rather than the amount of data sent in total.

What the solution is here, is perhaps less than ideal. `cat` will drain the fifo's read end successfully with no drama whatsoever. I can only assume that this is due to the way that it reads a file (blocking `read` calls in a loop) vs using `libdispatch`. I discovered this when debugging fifo behaviour using `--stdout=` within `simctl`. It appeas that if you use fifos here for stdout/stderr and instead use `cat` to drain them in other processes that you will be fine. As such we can replicate this behaviour within `FBControlCore` by using `cat` as an intermediary to do "good reading" of a fifo's read end. If I ever get to the bottom of why this, we should be able to go back to the dispatch based approach.

As an aside, we're still using libdispatch here, as it's an implementation detail of how we read the output of tasks.

Reviewed By: c-ryan747

Differential Revision: D22188122

fbshipit-source-id: 54320097f13febc4e20ceec1d5033e2fa30d13d0