Split out to FBDeviceCommands
Summary:
The intent here is that we can have failing implementations of these when a device has either:
1) A restorable device and therefore unable to do much
2) Has lost connectivity and therefore the AMDevice is invalid.

This will aid with failing fast. By pushing this into a protocol, we can wrap all kinds of devices in a `FBDevice` instance but susbitute out implementations under the hood.

At the moment this purely extracts the protocol, there's no change in any of the callers right now.

Reviewed By: c-ryan747

Differential Revision: D22162213

fbshipit-source-id: 853f11335440ecde77dff4ac0b7226dd4bbecb25