Add option to switch over to the newer nested format
Summary:
This adds the option to switch to the newer format. This will work with older clients since this requires strict matching on the new format, if the value is undefined then it is the legacy format.

There's an additional API on the executor that's also no longer needed so this has been deleted

Reviewed By: c-ryan747

Differential Revision: D22256035

fbshipit-source-id: 0e2f64c6970f695f14b1f592f0555b73e69e2081