Add support for CoreSimulator based Accessibility
Summary:
Now that the `SimulatorBridge` is dying, we need to add the appropriate behaviour for the replacement.

This is `CoreSimulator` and `AccessibilityPlatformtranslation`-backed. This does exist before Xcode 12, but now it is essential. Fortunately (once you get your head around the API) it's far more feature rich than we had before:

- We have all the methods on the `NSAccessibility` protocol at our disposal!
- Hierarchies now exist
- It has less setup and appears less failure prone.

Reviewed By: c-ryan747

Differential Revision: D22209181

fbshipit-source-id: a661972b164691e15340c0f4ae634e1dbfd5ccee