Add H264 Streams to FBSimulatorBitmapStream
Summary:
The availability and reliability of video recoring in `simctl` varies a lot by iOS and Xcode version. It also now seems that the fragmented mp4 output is gone for good.

The bitmap streaming before is still here, so I've chosen to use `VideoToolbox` to do the encoding, whilst retaining full control over the output. The output here is a AnnexB .h264 file. This means that there's no AAVC container, but this can be fixed by using ffmpeg. As such having a raw H264 stream is significantly more adaptable than any format. What's more is that the output is done in real time, on a hardware video encoder. This drastically reduces the amount of memory and CPU required to get usable encoded video out of a Simulator.

Containerization can be done downstream here and it's easy to add this to a video pipeline. The stream based approach is also useful for idb. Files can be streamed to on the remote side, or even over the network if the stream is pushed over grpc.

Reviewed By: c-ryan747

Differential Revision: D22283296

fbshipit-source-id: 56a80092b0c7bf9d5f43c7b43188f89e0c4bf839