FBDeviceSet composes from both FBAMDeviceManager and FBAMRestorableDeviceManager
Summary:
This will allow us to substitute a device for a restorable device when this state transition occurs. This is done by making `FBDeviceSet` the delegate of both the `FBAMDeviceManager` and the `FBAMRestorableDeviceManager`.

With this in place we'll get the appropriate callbacks at all times and we can make the appropriate dispatches. A Device is now "existing" when either a `FBAMDevice` or an `FBAMRestorableDevice` for it exist. Once both of these dissapear, the device goes bye-bye.

This will become useful as we add *overlapping* of restorable and regular devices. For the time being, in practice the device is a XOR of restorable or regular device. A future diff will make sure that there can be both kinds of backings.

Reviewed By: fejese

Differential Revision: D22389212

fbshipit-source-id: e53610aca034e2b8d28dd34682842cca29420639