ECID Based filtering
Summary:
This will aid in over-fetching for simple queries. ECID implies Devices, so we can just use the `--only` param and detect this.

The filter in this case works down at the detection layer. This means we plain don't inflate any information about a device if it doesn't match the filter. As a result this can result in faster returns with a large number of devices connected, since we're not overfetching

Reviewed By: fejese

Differential Revision: D22431810

fbshipit-source-id: 61502d4e7231b1feaf77bb69ce516ca2864d8da9