FBDevice is a facade over FBAMDevice &| FBAMRestorableDevice
Summary:
Modify the `FBDevice` class to be backed by either or both a `FBAMDevice`/`FBAMRestorableDevice`. Doing this means that we can substitute between the two as devices enter different states.

The `FBAMDevice` is always preferred over an `FBAMRestorableDevice`, this is completely by design as it means that we get more information and implementations for many commands. If a device becomes "restorable only", then we will default to using the base implementation (in D22435967) which will always fail, but fail gracefully

Reviewed By: fejese

Differential Revision: D22435968

fbshipit-source-id: 38a3c5c86152c215cfb1a048cd9f276b89a70e8b