Add AMDeviceNotificationTypePaired
Summary:
I've not seen this anywhere else, but as best I can tell, the notification value of 4/0x4 represents a pairing of a device.

I can see this in `__AMDDevicePaired`, which will log tha the device has entered a pairing state. This then calls the callback with `0x4` as the notification.

Reviewed By: fejese

Differential Revision: D22453520

fbshipit-source-id: 948c9b732b516dfbb8171b1d6d36ac1838009904