Add terrible hack to deal with async restorable device population
Summary:
This is an unfortunate hack to deal with the fact that a restorable device will be populated asynchronously from the main thread.

This is because the restorable device callbacks are delivered on a background queue, by which time the main thread has advanced. This means that there is a race condition where the first device callback is delivered after querying the device set for devices. As a result we can miss them, even though the logging indicates that a device is found after the failure

Reviewed By: fejese

Differential Revision: D22944254

fbshipit-source-id: bbb1e3d5b4bc2fff5d3015db142ae820965f4596