Don't expose FBAMDevice
Summary:
This will allow us to completely hide the `FBAMDevice` class, since we have a protocol in `FBDevice` (the protocol).

Doing this is important along, where I plan to add references for restorable devices. Hiding this behind the cluster of `FBAMDevice` and `FBRestorableDevice` will then allow us to switch to the appropriate implementation depending on the incoming command

Reviewed By: fejese

Differential Revision: D22943190

fbshipit-source-id: fb0ef2815cb9af76f733c30f9625ca236b6b322b