Add -[FBDeviceRecoveryCommands exitRecovery]
Summary:
This will correspondingly only operate on devices in the recovery state. This is fine as we can handle restorable (and not AMDevice) backed devices.

In these cases we need to extract an additional part of the device cluster, this is the recovery mode device. This is only present in restore mode devices. Then we just need to issue the right set of calls to make this happen.

Reviewed By: fejese

Differential Revision: D22943864

fbshipit-source-id: e0eafd690629e982cddff067320ee802638741d4