Set Secure Context (if present)
Summary:
Something has changed in newer iOS Version where the crash report service now uses a "Secure Context"

Fortunately this isn't too tricky to use, we just need to propogate it. This only occurs in the case that we create a connection ourselves. In this case we just need to check the secure context of the connection and then apply it to the AFCConnection

Reviewed By: fejese

Differential Revision: D22999636

fbshipit-source-id: 30352ebdc812717087c63442b37620db0b6112ba