fileCommandsForMediaDirectory
Summary:
This clears up ambiguity around "root" in the context of file commands.

For example "root" doesn't actually mean root on devices, it's the media directory. Based upon this we have to re-name internally, but keep compatibility with older clients. For devices "root" means "media" for sims "root" means "root".

This will be further disambiguated later on

Reviewed By: c-ryan747

Differential Revision: D23026591

fbshipit-source-id: 2acfe3f721c2f0314b510171e459116a608fa243