Don't populate AMDevices from list
Summary:
This brings listing in line with `FBAMRestorableDevice`.

This also stops the device manager from getting the same device multiple times on launch. This is completely intentional, there's a slight increase in latency for the device to show up but we still wait before repoting for a similar reason to Restorable Devices; they are delivered asynchronously in callback mechanisms.

Reviewed By: zeyadsalloum

Differential Revision: D22973891

fbshipit-source-id: 0628df589370944ab1e18af28490b2357cf990e2